[awesome] fix cpu usage graph updates
Now the graph actually shows the usage trend over time rather than just
spikes of cpu usage (which wasn't referring to actual usage).  What
seems to have been required here is a wait time of at least 3 seconds so
that the data is collected and can be plotted.  It could be that
previously the data was coming in too fast for the plot.  A wait time of
5 seconds also works, however 2 seconds creates the incorrect behaviour
again, hence the lowest working time has been chosen.