[vim] fix dark grey backouground for cursorline
Now that we're using alacritty instead of gnome-terminal, we now have a
"real" black as the background.  Hence, the hack which used Gnome's
black to fake a dark grey background no longer works.  Using a specific
number from the xterm256 colour set now works and is a much better fit,
and makes more logical sense.