[mailcap] fix firefox not finding mutt.html files
There was the problem that the first time a page was loaded from mutt,
that firefox couldn't find the file /tmp/mutt.html.  Now, with the
copiousoutput option, firefox finds the files correctly.  How one is
supposed to know this I don't know, but I managed to spot the solution
on stack overflow (noted in a comment).