[vim] mention that neon colorscheme needs vim-scripts package
While setting up a new system I was wondering how I came to install this
colour scheme when it didn't seem to exist.  Only after having installed
`vim-pathogen` did the colorscheme exist.  I found out that the colour
scheme (and many more) are available as part of the `vim-scripts`
package which is an upstream dependency of `vim-pathogen` (on Debian).