[bash] Avoid keyring when installing Python packages via poetry
The Python `poetry` tool likes to try to use the keyring to find
passwords when it adds and/or installs Python packages.  I don't use a
keyring, hence the popup that this shows is annoying and a waste of
time.  To avoid this problem, we need to use the "null keyring", which
is set here via the `PYTHON_KEYRING_BACKEND` environment variable.